bau: Use docker in pre-commit
I was pretty surprised to see the tests in this (java) project attempt
to download a bunch of python and JS dependencies. Needless to say, my
host OS is not set up with the exact set of versions to install these
deps, so the build fails. Seems like this would be a good use of docker.

This commit changes pre-commit to run in docker by default (you can
still run without by calling gradle directly or by passing in
--no-docker).

On Jenkins the working directory is mounted as a volume, which means
docker is able to change files on the host. It feels more docker-y to
build everything into the container, so I've updated the dockerfile to
do that. This does means you need to rebuild the container when you run
the tests though.

I'm copying the build.gradle across in a separate step so that the
dependencies are cached in a separate layer to the source.

Finally, I've collapsed as many of the commands in the dockerfile as possible
into one to avoid creating too many layers.