TRELLO-W0VXgdcb: Hash idp entity id and pid before storing them in db
Audit and Billing applications hash an event details' identity provider
entity id and its persistent id to create a hashed persistent id and
override the event details' persistent id with the hashed persistent id.
The event is then stored in the database so that the audit and billing
reporting systems can use the hashed persistent id to produce audit and
billing reports. This change ensures that Event Emitter performs in the
same way as Audit and Billing applications so that event details in
PostgreSQL are same as the event details in MongoDB.

Authors: @adityapahuja