Add encrypted event when logging errors
As the Event Sink MongoDB no longer exists, the only way we will be able to replay messages which failed to send is by capturing them in Kibana.  This is intended to be a stop-gap solution; in the long run we think it would be preferable for user journeys to fail if the audit logger is unavailable.

https://trello.com/c/y8ieTzSO/14-needs-kick-off-adjust-event-emitter-to-log-encrypted-event

Solo: @alan-gds