HUB-908: Fix integration tests
The version of localstack we were using was a little out of date and not
working. This bumps to the latest versions and configures it correctly.

The error message in the second integration test now returns a random
request ID, so the expected string has been shortened. The test
frameworks checks that the expected string is a substring of the actual
message, so this still works and checks what we actually care about.