BAU: Bump dropwizard-logstash and common-test-utils
This is being done as they both now use dropwizard 2, and we want to
upgrade the hub.

This seems like a pointless commit at the moment as they versions are
set to SNAPSHOT. This commit will get updated before merging and once
those versions exist (waiting on PRs). This serves as a reminder.